--- culture-ticketing-api-server/src/main/java/com/culture/ticketing/notification/infra/EmitterRepositoryImpl.java
@@ -1,49 +1,0 @@
-package com.culture.ticketing.notification.infra;
-
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
-
-public class EmitterRepositoryImpl implements EmitterRepository {
-
-    private final Map<String, SseEmitter> emitters = new ConcurrentHashMap<>();
-    private final Map<String, Object> eventCache = new ConcurrentHashMap<>();
-
-    @Override
-    public SseEmitter save(String emitterId, SseEmitter sseEmitter) {
-
-        emitters.put(emitterId, sseEmitter);
-        return sseEmitter;
-    }
-
-    @Override
-    public void saveEventCache(String emitterId, Object event) {
-
-        eventCache.put(emitterId, event);
-    }
-
-    @Override
-    public void deleteById(String emitterId) {
-
-        emitters.remove(emitterId);
-    }
-
-    @Override
-    public Map<String, Object> findAllEventCacheStartWithById(String userId) {
-
-        return eventCache.entrySet().stream()
-                .filter(entry -> entry.getKey().startsWith(userId))
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    }
-
-    @Override
-    public Map<String, SseEmitter> findAllEmitterStartWithById(String userId) {
-
-        return emitters.entrySet().stream()
-                .filter(entry -> entry.getKey().startsWith(userId))
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    }
-
-}
